Working on stereo checkerboard detector.  Adding pose fitting helper code for stereo checkerboard detection.
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@15463 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: 7a69e84d0bc8ca4a69e0fb673f8dfc06dac1ffc7

+++ vision/stereo_checkerboard_detector/src/reprojection_helper.cpp
@@ -95,8 +95,8 @@ void ReprojectionHelper::reproject(const std::vector<robot_msgs::Point>& ros_uvd
 {
   const unsigned int N = ros_uvd.size() ;
 
-  CvMat* uvd = cvCreateMat(N, 1, CV_64FC3) ;
-  CvMat* xyz = cvCreateMat(N, 1, CV_64FC3) ;
+  CvMat* uvd = cvCreateMat(N, 1, CV_32FC3) ;
+  CvMat* xyz = cvCreateMat(N, 1, CV_32FC3) ;
 
   for (unsigned int i=0; i<N; i++)
   {
